+++ docs/README.md
@@ -1,0 +1,35 @@
+# 기능 구현 목록
+
+## 도메인
+
+### Category
+
+### Couch
+
+### DataRow
+
+### DataRows
+
+### Day
+
+### Menu
+
+## 뷰
+### InputView
+### OutputView
+
+## 서비스
+
+### MenuController
+
+### MenuRepository
+
+### Validator
+- 주어진 입력의 유효성 검증
+- 코치는 적어도 2명 이상!
+- 음식 이름은 유효한 이름만 가능
+
+
+
+
+
