+++ src/main/java/menu/Application.java
@@ -1,5 +1,7 @@
 package menu;
 
+import menu.service.MenuController;
+
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
